MDEV-21117: --tc-heuristic-recover=rollback is not replication safe
Problem:
=======
When run after master server crash --tc-heuristic-recover=rollback produces
inconsistent server state with binlog still containing transactions that were
rolled back by the option. Such way recovered server may not be used for
replication.

Fix:
===
A new --tc-heuristic-recover=BINLOG_TRUNCATE value is added to the
server option.  It works as ROLLBACK for engines but makes its best to
synchronize binlog to truncate it to the last committed transaction.

The BINLOG_TRUNCATE algorithm starts scanning possibly a sequences of
binlog files starting from the binlog checkpoint file to find a
truncate position candidate.  The candidate truncate position
corresponds to the beginning of the first GTID event whose transaction
is not found to have been committed in any of involved engines should
they be multiple.
There must be no committed
transactions beyond this position *but* non-transactional events are
possible. This is also checked to produce an error when the case, and
no changes are done to transactions-in-doubt.
In the safe (expected) case possible partially committed transactions
get fully committed and
the binlog GTID state is adjusted to hold the GTID of the last
committed transaction.
Finally a STOP binlog event is written at the end of
truncated binary log.